move dsl-compile file
