+++ sdk/python/setup.py
@@ -89,7 +89,7 @@ setuptools.setup(
     include_package_data=True,
     entry_points={
         'console_scripts': [
-            'dsl-compile = kfp.compiler.main:main',
+            'dsl-compile = kfp.cli.compile:main',
             'dsl-compile-deprecated = kfp.deprecated.compiler.main:main',
             'kfp=kfp.cli.cli:main',
         ]
